refractored my methods with the new code
